resend activation page updated
