+++ FrontEnd/src/pages/SignUp/SignUp/ResendActivationPage.jsx
@@ -1,29 +1,52 @@
+import { useState } from 'react';
 import { Link } from 'react-router-dom';
-import styles from './SignUpPage.module.css';
-import { ResendActivationFormComponent } from '../ResendActivationForm';
+import { ResendActivationFormContent } from '../SignupForm/ResendActivationFormContent.jsx';
+import styles from './ResendActivationPage.module.css';
 
 export function ResendActivationPage() {
+  const [isValid, setIsValid] = useState(false);
+
+  const updateIsValid = (value) => {
+    setIsValid(value);
+  };
+
   return (
-    <div className={styles['sign-up']}>
-      <div className={styles['sign-up__body']}>
-      <img className={styles['sign-up-img-right']}
-          src={`${process.env.REACT_APP_PUBLIC_URL}/img/dots_7x6.png`}
-          alt="dots_7x6.png" />
-        <div className={styles.container}>
-          <ResendActivationFormComponent />
-          <div className={styles['sign-in-line']}>
-            <div className={styles['sign-in-line__text']}>
-              Вже були на нашому сайті?
+    <div className={styles['resend-activation__page']}>
+      <div className={styles['resend-activation__body']}>
+        <div className={styles['resend-activation__container']}>
+            <div className={styles['form__header']}>
+              Надіслати лист для активації ще раз
             </div>
-            <Link to="/login" className={styles['sign-in-line__link']}>
-              Увійти
-            </Link>
+            <div className={styles['form__content']}>
+                <p className={styles['form__text']}>
+                  Введіть електронну адресу вказану при реєстрації для повторного
+                  надіслення листа. <br />
+                  На зазначену Вами електронну пошту буде відправлено листа з посиланням
+                  для активації. <br />
+                </p>
+                <ResendActivationFormContent setIsValid={updateIsValid} />
+                </div>
+            <div className={styles['form__footer']}>
+              <div className={styles['button-container']}>
+                <button
+                  disabled={!isValid}
+                  form="signUpForm"
+                  className={
+                    isValid
+                      ? styles['resend-activation__button']
+                      : styles['resend-activation__button__disabled']
+                  }
+                  type="submit"
+                >
+                  Надіслати
+                </button>
+                <Link className={styles['cancel__button']} to="/sign-up">
+                  Скасувати
+                </Link>
+              </div>
+            </div>
           </div>
         </div>
-        <img className={styles['sign-up-img-left']}
-          src={`${process.env.REACT_APP_PUBLIC_URL}/img/dots_7x6.png`}
-          alt="dots_7x6.png" />
-      </div>
     </div>
   );
 }
