+++ FrontEnd/src/pages/SignUp/SignUp/ResendActivationPage.module.css
@@ -1,0 +1,144 @@
+.resend-activation__page {
+    width: 100vw;
+    border-radius: 8px;
+    background: var(--light-seashell-background);
+    padding: 80px 0;
+    display: flex;
+    justify-content: center;
+}
+
+.resend-activation__body {
+    display: flex;
+    align-self: center;
+    flex-direction: column;
+    align-items: center;
+    max-width: 572px;
+    min-width: 345px;
+    gap: 24px;
+}
+
+.resend-activation__container {
+    display: flex;
+    width: 375px;
+    flex-direction: column;
+    align-items: flex-start;
+    border-radius: 6px;
+    background: var(--main-white, #fff);
+}
+
+.form__header {
+    display: flex;
+    align-self: stretch;
+    padding: 16px 24px;
+    box-shadow: 0px -1px 0px 0px #f0f0f0 inset;
+    color: var(--character-title-85, rgba(0, 0, 0, 0.85));
+    font-feature-settings: "calt" off;
+    font-family: var(--font-main);
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: 20px;
+    letter-spacing: -0.16px;
+}
+
+.form__footer {
+    display: flex;
+    padding: 16px 24px;
+    align-items: center;
+    gap: 8px;
+    align-self: stretch;
+    background: var(--main-white, #fff);
+
+    /* border & divider/divider ↑ */
+    box-shadow: 0px 1px 0px 0px #f0f0f0 inset;
+}
+
+.form__content {
+    display: flex;
+    flex-direction: column;
+    gap: 24px;
+    padding: 24px;
+}
+
+.form__text {
+    font-family: var(--font-main);
+    font-size: 14px;
+    font-weight: 400;
+    line-height: 22px;
+    letter-spacing: -0.01em;
+    text-align: left;
+    color: var(--main-text-color);
+}
+
+.button-container {
+    display: flex;
+    align-items: flex-start;
+    gap: 12px;
+}
+
+.cancel__button {
+    display: flex;
+    padding: 9px 16px;
+    border-radius: 4px;
+    border: 1px solid #232424;
+    background: var(--main-white, #fff);
+    color: #232424;
+    text-align: center;
+    font-feature-settings: "calt" off;
+    font-family: var(--font-main);
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: 20px;
+    letter-spacing: -0.16px;
+    text-decoration: none;
+    cursor: pointer;
+}
+
+.resend-activation__button,
+.resend-activation__button__disabled {
+    padding: 9px 16px;
+    justify-content: center;
+    align-items: center;
+    gap: 4px;
+    border-radius: 4px;
+    background: #000000;
+
+    color: var(--light-button-text-color);
+    text-align: center;
+    font-feature-settings: "calt" off;
+
+    font-family: var(--font-main);
+    font-size: 16px;
+    font-weight: 600;
+    line-height: 20px;
+    letter-spacing: -0.16px;
+
+    cursor: pointer;
+}
+
+.resend-activation__button__disabled {
+    border: 1px solid #BFC6CF;
+    background: #EEEFF1;
+    color: #A3AAB0;
+}
+
+@media only screen and (min-width: 768px) {
+    .resend-activation__page {
+        padding: 104px 0;
+    }
+
+    .resend-activation__container {
+        width: 572px;
+    }
+
+    .form__content {
+        gap: 10px;
+    }
+}
+
+@media only screen and (min-width: 1512px) {
+    .resend-activation__page {
+        padding: 120px 0;
+    }
+}
