+++ FrontEnd/src/pages/SignUp/SignupForm/ResendActivationFormContent.module.css
@@ -1,9 +1,8 @@
 .resend-activation-form {
     display: flex;
-    padding: 24px;
     flex-direction: column;
     align-items: flex-start;
-    gap: 10px;
+    gap: 4px;
     align-self: stretch;
 }
 
@@ -11,22 +10,21 @@
     display: flex;
     flex-direction: column;
     align-items: flex-start;
+    align-self: stretch;
     gap: 8px;
+    width: 100%;
 }
 
 .resend-activation-form__row {
-    display: flex;
-    align-items: flex-start;
-    gap: 8px;
+    width: 100%;
 }
 
 .resend-activation-form__column {
     display: flex;
-    width: 257px;
-    height: 84px;
     padding-bottom: 0px;
     flex-direction: column;
     align-items: flex-start;
+    width: 100%;
 }
 
 .resend-activation-form__label {
@@ -67,7 +65,6 @@
 
 .resend-activation-form__field {
     display: flex;
-    width: 520px;
     align-items: center;
     align-self: stretch;
     border-radius: 2px;
@@ -77,7 +74,7 @@
 .resend-activation-form__input {
     display: flex;
     height: 22px;
-    padding: 5px 12px;
+    padding: 9px 12px;
     align-items: center;
     gap: 4px;
     flex: 1 0 0;
@@ -87,11 +84,15 @@
     color: var(--character-title-85, rgba(0, 0, 0, 0.85));
 
     /* Body/regular */
-    font-family: var(--font-messages);
+    font-family: var(--font-main);
     font-size: 14px;
     font-style: normal;
     font-weight: 400;
     line-height: 22px;
+}
+
+.resend-activation-form__input::placeholder {
+    color: var(--placeholder-text-color);
 }
 
 .resend-activation-form__input:focus {
@@ -104,12 +105,10 @@
     align-items: flex-start;
     gap: 10px;
     align-self: stretch;
-    color: var(--red-red-100, #f34444);
-
-    /* Body/regular */
-    font-family: var(--font-error);
-    font-size: 14px;
+    color: var(--error-text-color);
+    font-family: var(--font-main);
+    font-size: 12px;
     font-style: normal;
     font-weight: 400;
-    line-height: 22px;
+    line-height: 16px;
 }
